Add working RuboCop defaults
Replaces `.rubocop.yml` (which wasn’t working with vanilla RuboCop *or* the
Code Climate Rubocop engine) with a file which just inherits from
`.rubocop_defaults.yml`. The latter file contains the concatenated RuboCop
defaults (copied from the RuboCop repository), with several cops and options
commented out because they did not work with one or the other or both RuboCop
engines. The Rails cops are also enabled.

There’s probably a lot of duplication here (and it’s not in the least bit
maintainable), but an elegant RuboCop configuration is a job for another day
(maybe once we’ve fixed the security vulnerabilities first...).